backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic>=2.0
google-generativeai
aiofiles==23.2.1
openpyxl
pandas
slowapi==0.1.9
httpx==0.25.2
